Moved the daemon code into the TLD
