+++ common/settings.go
@@ -1,0 +1,8 @@
+package swcommon
+
+import "os"
+
+const (
+	fileMode os.FileMode = 0600
+	dirMode  os.FileMode = 0700
+)
