+++ go.mod
@@ -10,7 +10,7 @@ require (
 	github.com/goccy/go-json v0.10.2
 	github.com/google/uuid v1.6.0
 	github.com/gorilla/mux v1.8.1
-	github.com/nodeset-org/hyperdrive v0.5.0-dev.0.20240415155425-3de675347f68
+	github.com/nodeset-org/hyperdrive-daemon v0.0.0-20240415231824-bdb6041e2a20
 	github.com/rocket-pool/batch-query v1.0.0
 	github.com/rocket-pool/node-manager-core v0.2.1-0.20240414212301-586ea27e9ef4
 	github.com/urfave/cli/v2 v2.27.1
