--- stakewise-daemon/common/settings.go
@@ -1,8 +1,0 @@
-package swcommon
-
-import "os"
-
-const (
-	fileMode os.FileMode = 0600
-	dirMode  os.FileMode = 0700
-)
